HTTP/Headers: add Tk
The draft of the Tracking Preference Expression specification [1]
doesn't mention any browser behaviour linked to the `Tk` header.
Therefore it is hard to say whether a browser is compatible.

A browser could parse and visualize the content of the header,
but this should not be considered a requirement for compatibility.

[1] https://www.w3.org/2011/tracking-protection/drafts/tracking-dnt.html#Tk-header-defn