Remove deprecated flag from Navigator: vendorSub
This change updates the status data for Navigator: vendorSub to indicate
that it’s not deprecated. While it’s true that the property isn’t
reliably useful, that doesn’t mean it’s formally deprecated — it’s not
in fact, because the HTML spec defines normative requirements for it.

https://html.spec.whatwg.org/multipage/system-state.html#client-identification

So it’s misleading to mark it as deprecated.