Remove deprecated flag from some NavigatorID props (#2965)
This change updates the status data for several properties of
NavigatorID to indicate that they are not deprecated. While it’s true
that the properties aren’t reliably useful, that doesn’t mean they’re
formally deprecated — they’re not in fact, because normative
requirements for their values are defined in the HTML spec.

https://html.spec.whatwg.org/multipage/system-state.html#client-identification

So it’s misleading to mark them as deprecated.