IE11: inset must be last in box-shadow (#2996)
* IE11: inset must be last in box-shadow

When specifying the box-shadow property, the optional `inset` specification must be at the end. Otherwise, Internet Explorer will ignore the full property. I have only tested this on IE11, but I suspect earlier versions have the same problem. I have not found any documentation or questions about this quirk online. 

Aside: I have put the note in the same string as the original note. Is this the right thing to do? Or should it be converted to an array of two strings?

* Apply feedback

* Added comma