Add spec_url data for JavaScript features
This change adds spec URLs in the `*.json` sources for all JavaScript
features that have an `mdn_url` for an MDN article with a
**Specification(s)** table — with the exception that no spec data is
added for any cases where a URL found in an MDN **Specification(s)**
table has no fragment-ID part.

The new field that holds the spec-URL data is named `spec_url`, and its
value is either a single URL (in the case where the feature is only
associated with a single specification references), or any array of URLs
(in the case where the feature is associated with multiple
specification references).

The change also updates `schemas/compat-data.schema.json` to allow the
additional spec data and to specify the structure it must conform to.