Add missing Edge 13 data based on Confluence
Produced by `npm run confluence -- --fill-only --browsers=edge` with
local changes applied to only include changes where the new version
was 13. Since Edge 12 was the first version of Edge and we don't have
data for Edge 12, this means we can't know if these APIs were first in
Edge 12 or Edge 13, so it can only say true.