Updating intrinsic sizing keywords data
Firefox 66 has unprefixed the min-content and max-content keywords. I'veupdated the compat data based on this and my testing. The fit-content keyword is still left prefixed.

The data is pretty patchy so I also tested on some other browsers to fill out some nulls, so have updated Edge and Safari too.