Update Safari (macOS/iOS) AudioContext support (#3372)
* Update Safari (macOS/iOS) AudioContext support

Safari for both macOS and iOS support the `AudioContext` API with the `webkit` prefix.  While developing a visualizer for my website and testing it out on iOS, I found that all it simply needed was a `context.resume()` as a part of compliance for [Chrome's autoplay policy](https://developers.google.com/web/updates/2017/09/autoplay-policy-changes).  This compatibility table has been updated to include supported functions, as tested with the latest Safari versions.  (This also applies a quick fix to an Opera compatibility statement that was set to `false` when it should be `true`, as proven by my own use of the function.)

* Set output_latency to false for Safari iOS