Update the support status for `initAnimationEvent` (#3486)
* Update the support status for `initAnimationEvent`

`initAnimationEvent` was never supported in WebKit, confirmed by

`git log -S initAnimationEvent` in the WebKit source tree, where the
only cases are in a test from Chakra [1] and in spec drafts [2], not
in any code.

Also confirmed by testing various old versions of Chrome and Safari with:
https://software.hixie.ch/utilities/js/live-dom-viewer/?saved=6671
https://software.hixie.ch/utilities/js/live-dom-viewer/?saved=6672

A `initWebKitAnimationEvent` was supported, but it was removed in 2011:
https://trac.webkit.org/changeset/103751/webkit

Every WebKit- and Blink-based browser should therefore not support
this. For Opera, which did support it with Presto, set it to removed
in version 15, the first Chromium-based Opera release.

[1] https://trac.webkit.org/changeset/205387/webkit
[2] https://trac.webkit.org/changeset/66236/webkit

* Add data for `initWebKitAnimationEvent`

Using http://software.hixie.ch/utilities/js/live-dom-viewer/?saved=6723
on BrowserStack.

The original `version_added` dates don't seem reliable, it was confirmed
present on iOS 3 and might have been present earlier. For Chrome the
original data was about `initAnimationEvent` and wrong.