Set CSS properties to not supported based on `CSS.support` tests
These results are based on results collected from
https://github.com/foolip/mdn-bcd-collector, where for each CSS
property found in reffy-reports, a `CSS.supports(propertyName,
"initial")` was generated. The results are in:
https://github.com/foolip/mdn-bcd-results/tree/19c48eeb39943cf00c84da3f87e23d18e7e96334

The script used to apply those results to BCD is this:
https://gist.github.com/foolip/64ea409ba957adc54a756dd7936627d7

In the browsers affected by this change, the
`CSS.supports("display", "initial")` test was confirmed to return
true, making it unlikely that these results are incorrect.