update compatibility for window.parent
I reviewed this compatibility on browser stack for both IE and Safari. For validation I used `window.parent === window => true` meant it was implemented. Safari was valid as far back as BrowserStack provides which is Safari 4 so I set it to `true`. IE 8 evaluated to `false` but every version above that evaluated to `true` so I've set the supported version to 9.

If you'd like additional testing let me know.