Mark some CSS properties as first available in Safari 5 / 5.1
This is based on results collected from Safari 4, 5 and 5.1 (and more):
https://github.com/foolip/mdn-bcd-results/tree/6288e5f0813157d70821c635fe05bb469ee96e0d

There are no release notes for Safari 5 or 5.1, but these changes are
likely to be correct. Tests on the form
`'propertyName' in document.body.style` were used to determine support.

The script to update results based on resutls is in development:
https://gist.github.com/foolip/bdf30013e20342b35cb54d6f5f589372

These changes were hand picked from the output as good changes.