Change Chrome API support from null to true based on tests
Tests were generated by https://github.com/foolip/mdn-bcd-collector
and results collected in Chrome 73 on Android 9 and Windows 10:
https://github.com/foolip/mdn-bcd-results/pull/111
https://github.com/foolip/mdn-bcd-results/pull/112

This is similar to https://github.com/mdn/browser-compat-data/pull/3700,
which has examples of what the tests look like.

Where the results were true and the existing data in BCD was null, the
value was changed to true. Snapshot of script used:
https://gist.github.com/foolip/57fb835508ce96281854cc2ba44277ad

Plain false positives are very unlikely, but some of the exposed APIs
might still need notes or could be partial implementations.