Change Opera API support from null to true based on tests
Tests were generated by https://github.com/foolip/mdn-bcd-collector
and results collected in Opera 58 on Windows 10 and Opera 51 on Android
9:
https://github.com/foolip/mdn-bcd-results/pull/114
https://github.com/foolip/mdn-bcd-results/pull/117

This is similar to https://github.com/mdn/browser-compat-data/pull/3700,
which has examples of what the tests look like.

Where the results were true and the existing data in BCD was null, the
value was changed to true. Snapshot of script used:
https://gist.github.com/foolip/b733127722b283e552a580b5421ccc23

Plain false positives are very unlikely, but some of the exposed APIs
might still need notes or could be partial implementations.