Add estimated release date for IE 1.5
The Wikipedia article puts the release of IE 1.5 at January 1996, but doesn't specify a date.  I don't think we'll be able to narrow down the exact release date since it's so long ago, so I would say it's good enough?  I'd love to know if other people think we should try to find more accurate data.