Firefox 67 implements MediaDeviceInfo.groupId (#4093)
* Fix version for MediaDeviceInfo.groupId for Firefox
This was added in Firefox 67, not Firefox 39.

* Fix MediaDeviceInfo.groupId to be more detailed
Turns out that groupId was added in Firefox 39 but didn't actually
use the same value for related devices until 67. This commit
corrects the data to reflect that.

* Address review feedback and fix Opera info
This fixes review feedback from @darkvinylscratch to simplify
the Firefox info structure. This also fixes the Opera info to
not claim it's supported while at the same time having a note saying
you have to use a polyfill. Now it says it's unsupported, but available
with adapter.js. Also, added `<code>` around `toJSON()`.