Show support in Edge for @charset
Verified by testing in Edge 18, but I have no reason to believe this was ever unsupported given how old this feature is in the web platform. I also ran a few searches to check if anyone had complained about issues using this feature in Edge and did not find anything.