Update Navigator to note secure context needed for getUserMedia() (#4335)
The Media Capture and Streams spec has been updated to require the use
of a secure context -- one loaded using HTTPs, localhost, or file:/// --
in order to access navigator.mediaDevices.getUserMedia(). This updates
the data accordingly. Currently this is only enforced in Chrome. Firefox
will begin to do so soon, but is currently doing so on Nightly builds.

Sources:
* [Disable getUserMedia on non-secure origins](https://bugzilla.mozilla.org/show_bug.cgi?id=1335740)
* [intent to deprecate](https://groups.google.com/forum/#!topic/mozilla.dev.platform/vmO0NRM46l8)