Add Safari status for CSS clip-path (#4356)
* Add Safari status for CSS clip-path

Added to Safari as part of [bug 95474](https://bugs.webkit.org/show_bug.cgi?id=95474) and linked [changeset](https://trac.webkit.org/changeset/127327/webkit). Safari 7 seems likely to be the first major release, but it's a bit hard to be sure without a good archive of Safari versions.

* Add Safari unprefixed status for CSS clip-path

Some versions of Safari also support `clip-path` without a prefix, but it is
currently unknown precisely when that began.