Update chrome in console.timeLog() compat (#4442)
* Update chrome in console.timeLog() compat

As I just discovered this, I don't know exactly when it was added, but the bug report says it got fixed on 'Tue, Oct 2, 2018, 10:06 PM GMT+2'. So this looks like to be Chrome 70. Also tested this in Chrome on Android with the remove devices tool.
(see https://bugs.chromium.org/p/chromium/issues/detail?id=854474#c8)

* Updating version numbers timeLog

According to vinyldarkscratch