Update for link element's disabled attribute being added back (#4323)
* Update for link element's disabled attribute being added back
At some point, the link element's disabled attribute was removed
from the spec. It has since been restored but was not linked to
the stylesheet's disabled attribute.

In https://github.com/whatwg/html/issues/3840, the two are being
revised to reflect one another. This BCD update adds data to
indicate this.

* First bit of corrections for emilio

* Further fixes

* Apply Chrome information to other browsers, with minor rewording to be
inclusive of all Blink-based products. Replaced old and incomplete note
on the Android WebView entry.