Remove old paragraph from partial impl. section
Under `partial_implementation`, the first and second paragraphs are nearly
identical. This is because the second one was added to replace the first,
in order to improve the description of the value (the old text made it easy
to reverse the presumed meaning of the value). Somehow, the original
paragraph did not get removed when the rewrite was added. This PR deletes
the older version of the paragraph and tidies the new paragraph just a
little bit more.