Updates to DOMPoint*, DOMRect*, and DOMQuad (#4828)
* Updates to DOMPoint*, DOMRect*, and DOMQuad
Firefox 69 introduced several changes to these. In
addition, a number of the existing entries were either
missing or wrong, and several members that were
added in the past were not listed. This should bring
everything up to date.

For the changes made, the sources are below.

*For all of these, Edge is not available as the entire
Geometry Interfaces API is under consideration still.
For Safari, the API was almost completely implemented on
first addition in Safari 10.1, with some stuff added in
Safari 11.*

Worker support information:
* [Safari](https://trac.webkit.org/browser/webkit/branches/safari-603.1.30.0-branch/Source/WebCore/dom/DOMPoint.idl)
* [Chrome](https://codereview.chromium.org/2650583009)
* [Firefox](https://bugzilla.mozilla.org/show_bug.cgi?id=1420580)

Addition of DOMQuad:
* [Safari](https://trac.webkit.org/browser/webkit/branches/safari-603.1.30.0-branch/Source/WebCore/dom/DOMRect.idl)
* [Firefox](https://bugzilla.mozilla.org/show_bug.cgi?id=917755)

Addition of static constructors (fromPoint(), fromRect(), fromQuad()):
* Safari: Present from the beginning
* [Firefox](https://bugzilla.mozilla.org/show_bug.cgi?id=1558101)

Thankfully, it's easy to bop around the WebKit code
to isolate versions. I adore how well they've tagged
and set up branches and releases to make it easy
to identify releases that implement things.

* Updates per review feedback
* Safari iOS 10.1 isn't a thing; it's actually 10.3.
* Added information for Samsung and Opera/Opera Android
* Confirmed that some other things were right, like
DOMQuad's introduction in iOS 11.