minor edit - change "13" to "13.0.0"
I've now also created [a PR](https://github.com/mdn/browser-compat-data/pull/5069) to add v13 to the list of Nodejs releases.  Following the example there I've listed v13 as "13.0.0" though the browsers appear to just use the 'major' version number and frankly I'd prefer to do the same since it is actually v13.0.1 that is current and that I tested again.  So I believe I've done it right now, but I submit this was some hesitation...