Record `URL` feature support in Chrome versions back to 37
Tested all four of these features on a stock Android 5 Lollipop
emulator image, in both the Chrome browser and a WebView, both at
version 37.

It's possible these features go back to the introduction of the URL
class itself in Chrome 32, but I don't have a handy way to test
Chrome versions between 30 and 37.

On the other hand the `searchParams` feature is indeed missing,
consistent with the existing data that has it start at Chrome 51.