Fx71 MediaRecorder updates (#5133)
* MediaRecorder.mimeType behavior improved in Firefox 71
This patch updates the BCD for the MediaRecorder interface to note that
the mimeType property has been updated to operate per the spec, with more
reliability.

Source:
* https://bugzilla.mozilla.org/show_bug.cgi?id=1512175

* Fx 71 adds audioBitsPerSecond/videoBitsPerSecond to MediaRecorder

* warning event and onwarning removed in Firefox 71

Technicall, the `warning` event was removed earlier, leaving
behind the `onwarning` event handler, which was not used, but I
have yet to find out when exactly `warning` went away. But `onwarning`
definitely was removed in 71.