Add Safari support version for plain href in textPath
WebKit 607.1 introduced support for the plain href attribute in textPath SVG
elements with the following patch:
https://trac.webkit.org/changeset/234683/webkit

https://trac.webkit.org/browser/webkit/trunk/LayoutTests/svg/custom/href-svg-namespace-elements.html?rev=234683#L16

Currently we declare that Safari supports this without specifying the version
number, this commit adds the support version information.

A manual test on BrowserStack confirmed that Safari 12.1 and iOS 12.2 both
support the plain href attribute in textPath SVG elements.