Mark FileSystemEntry as standard
`FileSystemEntry` is a core feature of the _File and Directory Entries
API_ https://wicg.github.io/entries-api/, an actively-maintained spec
https://github.com/WICG/entries-api that’s supported in Firefox as well
as in Chrome. So it should not be marked as non-standard.

However, the copyTo, moveTo, remove, getMetadata, and toURL members of
FileSystemEntry are obsolete and no longer in the spec — and so this
change marks them as deprecated.