Revert "Fix Chromium versions for font-size-adjust (#4762)"
This reverts commit bc674e2a51cb0c3de4aea8a1cd7bce71bcd41dd2.

The Pull Request for bc674e2 was raised based on a comment stating that font-size-adjust was enabled by default in Chrome 54. I've not been able to find any evidence of this.

According to Chrome Platform Status [1], Chromium Issue 451346 [2], Can I Use [3] and manual testing [4], font-size-adjust is still (as of Chrome 80) behind a flag.

[1]: https://www.chromestatus.com/features#font-size-adjust
[2]: https://bugs.chromium.org/p/chromium/issues/detail?id=451346
[3]: https://github.com/Fyrd/caniuse/blob/3588a7a15b4e5a3e7eab73eeb03ce22c5da68000/features-json/font-size-adjust.json#L165-L202
[4]: https://github.com/mdn/browser-compat-data/pull/4762#issuecomment-590781481

# Conflicts:
#	css/properties/font-size-adjust.json