Add implicit description to setLocal/RemoteDescription (#5814)
* mozGetAsFile() removed in Firefox 74
Updated data for `mozGetAsFile()` to indicate that it was
deprecated in Firefox 26 and removed in 74.

Source:
* https://bugzilla.mozilla.org/show_bug.cgi?id=1588980
* https://www.fxsitecompat.dev/en-CA/docs/2013/htmlcanvaselement-mozgetasfile-has-been-deprecated/

* Revert "mozGetAsFile() removed in Firefox 74"

This reverts commit 90728eeaff851a284cbae310baade17f911af2ab.

* Add implicit description to setLocal/RemoteDescription
Firefox 75 and Chrome 80 now support setLocalDescription()
and setRemoteDescription() without parameters; in this case,
the browser figures out and automatically creates an offer
or answer for you based on the current signaling state.

Sources:
* https://bugzilla.mozilla.org/show_bug.cgi?id=1568292
* https://bugs.chromium.org/p/chromium/issues/list?q=setLocalDescription%20argument&can=1

* Apply suggestions from code review

Accepted all suggestions. Even the "currently" one, after discussing it with ddbeck and despite still thinking it's fine, personally, and he conceding it's not a big deal. I guess I'm just being nice today. :)

Co-Authored-By: Daniel D. Beck <daniel@ddbeck.com>

Co-authored-by: Daniel D. Beck <daniel@ddbeck.com>