Update DataTransferItemList Chrome support data (#6074)
https://trac.webkit.org/changeset/98821/webkit is when
`DataTransferItemList` landed in WebKit. That puts it in
https://trac.webkit.org/log/webkit/tags/Safari-534.52.7,
which means it shipped in Chrome 13.

https://source.chromium.org/chromium/chromium/src/+/9f23921 and
https://src.chromium.org/viewvc/blink?revision=156015&view=revision show
that `DataTransferItemList.p.remove()` support was added to the Blink
sources in 2013-08-13. That was too late to make the Chrome 30 feature
freeze, which was on 2013-07-26. The next feature freeze was 2013-09-06,
for Chrome 31. So it seems this shipped in Chrome 31.