Fixed inconsistency in activeVRDisplays for Google
This had indicated version 79 for `ssecure_context_required`,
but since the version flag is `false` for the main feature,
this is not valid. So I've set these back to false and put
back in a note on the mobile Android entries about it requiring
a secure context.

Now I know why I did it this way in the first place. :D