Fix data for allowFullscreen on HTMLIFrameElement
This is messy because there's also a content attribute, and prefixed
variants of both, but the content attributes and IDL attributes didn't
always change at the same time. No webkit-prefixed variant of the IDL
attribute was ever exposed, and the moz-prefixed IDL attribute had a
different capitalization.

Version numbers found by bisecting on BrowserStack with this test:
https://software.hixie.ch/utilities/js/live-dom-viewer/saved/8300

The commits that added the final unprefixed form in each engine:
https://chromium.googlesource.com/chromium/src/+/9206c8a832a4ead0c9129fc110727f9155228d0c
https://hg.mozilla.org/mozilla-central/rev/91d78336d372
https://trac.webkit.org/changeset/205686/webkit

Because these changes are in bindings only, it's safe to assume that
these were supported at the same time in different browsers using the
same engine, including desktop/mobile.

The miscapitalized variant "allowfullscreen" that Gecko briefly had was
not added to the data.