Update support data for Animation and KeyframeEffect
The AnimationEffect data is already good, but Animation's effect
attribute and KeyframeEffect with constructor and target attribute
weren't updated.

These changes landed together in Chromium M75:
https://chromium-review.googlesource.com/c/chromium/src/+/1564568

The Animation and KeyframeEffect interfaces were verified to not exist
in Edge 18, so the Edge version was set/updated to "75". These data
points support this as well:
https://caniuse.com/#feat=web-animation
https://developer.microsoft.com/en-us/microsoft-edge/status/webanimationsjavascriptapi/