Fix the Safari versions for api.Response.clone
See https://github.com/mdn/browser-compat-data/pull/6515#issuecomment-676752274
for how this was tested in BrowserStack and Sauce, for macOS only.

Additional evidence is https://trac.webkit.org/changeset/197049/webkit,
where the whole Response interface was added, controlled by a single
flag. For the attributes/methods included in that commit, they were
almost certainly first shipped in the same version of Safari that
shipped Response and the overall Fetch API. This was 10.1 / 10.3 per
other existing compat data:
https://wiki.developer.mozilla.org/en-US/docs/Web/API/WindowOrWorkerGlobalScope/fetch#Browser_compatibility
https://developer.mozilla.org/en-US/docs/Web/API/Request#Browser_compatibility
https://developer.mozilla.org/en-US/docs/Web/API/Response#Browser_compatibility

There's still a lot clearly wrong with the data for Safari, but this
fixes one small piece of it.