Update versions for the initial WebKit implementation of Request
This is fixing issues similar to these recent PRs:
https://github.com/mdn/browser-compat-data/pull/6544
https://github.com/mdn/browser-compat-data/pull/6574

Fix the versions for those parts that were in WebKit's initial
implementation:
https://trac.webkit.org/browser/webkit/trunk/Source/WebCore/Modules/fetch/FetchRequest.idl?rev=195954

That the updated members were present in Safari 10.1 for macOS and
Safari for iOS 10.3 was confirmed with an ad-hoc test in BrowserStack:
https://software.hixie.ch/utilities/js/live-dom-viewer/?saved=8385

Note that `keepalive` wasn't in the initial implementation and added
some time later. It's not updated in this change for ease of review.