Update API compatibility for Firefox for Android 79
With Firefox for Android 79, we're building on a new browser engine. As
such, many WebExtensions APIs needed to be re-implemented from scratch.
With the release of 79 we haven't yet achieved WebExtension API parity,
therefore this changeset is updating API compatibility. We are working
on restoring the remaining APIs. Note also we have added a few APIs that
were not supported before.

Co-authored-by: Agi Sferro <agi@sferro.dev>