Update SVG <discard> support data (only implementation removed)
This was removed in Chromium 81: https://github.com/mdn/browser-compat-data/pull/5779

Using a simple `window.SVGDiscardElement` test, I determined that the
interface object was added in Chrome 34. It's possible that the markup
was supported earlier than that, but it's not important.

Since it's not possible for anyone to depend on this feature and the
data in BCD should be removed in ~2 years, just delete the subfeatures
which had no useful data, rather than researching the attribute support
in similar detail.