Remove irrelevant bugfix note
The bugfix in the note tracked unprefixed WebSocket usage, and was closed
in FF 11. Maintaining a link to it on the most recent version implies
the bugfix page is still relevant for readers, which is untrue.

Even if devs were targeting those older FF versions, the bugfix page
didn't contain anything they'd need to be aware of.