Mark NavigatorID members standard; most deprecated
This change consistently marks all members of NavigatorID as standard —
because they are all in fact defined in the HTML spec. And it marks
almost all as deprecated — because almost all of them are required to
return either a constant (and so, useless) value, or else are defined to
*optionally* return a constant value, but are strongly encouraged by the
spec to do return a constant value.

The only exceptions are:

* navigator.userAgent, defined to return a non-constant value

* navigator.vendor, defined to return a value depending on which engine
  the browser is running; either "Google Inc.", "Apple Computer, Inc.",
  or the empty string (for Gecko)
  https://html.spec.whatwg.org/multipage/#dom-navigator-useragent

Fixes https://github.com/mdn/browser-compat-data/issues/6567