Update audioTracks/videoTracks/textTracks versions
This change corrects the Safari version data for the audioTracks, videoTracks,
and textTracks members of the SourceBuffer interface to reflect, per
https://bugs.webkit.org/show_bug.cgi?id=123374, the actual version of
Safari which first shipped support for them.

The change also updates the Safari version data for the audioTracks,
videoTracks, and textTracks members of the HTMLMediaElement interface.

Relates to https://github.com/mdn/browser-compat-data/issues/6102