Update when getUserMedia() shipped in Chromium
navigator.mediaDevices.getUserMedia() and navigator.getUserMedia() were
enabled together in M53:
https://chromium.googlesource.com/chromium/src/+/d229d009b74255797c16bef6b7e398a165dbeacd

The data for navigator.getUserMedia() was already correct, except for
Opera where it incorrectly appeared to still be prefixed-only.

The change to require a secure context for these APIs was made long
before it was shipped:
https://chromium.googlesource.com/chromium/src/+/dac67a137b9c8d3be8d51bb9fe8ab761d8b054a7

The versions for api.MediaDevices.getUserMedia.secure_context_required
should thus be the same as for the parent feature.