Update Safari data for visibilityState
This change updates the Safari data for visibilityState in light of
related changes made to the visibilitychange data in
https://github.com/mdn/browser-compat-data/pull/6849.

Specifically, testing indicates the following note isn’t accurate:

> Doesn't fire the visibilitychange event when document.visibilityState
> transitions to hidden

But what’s instead accurate is the following:

> Doesn't fire the visibilitychange event when navigating way from a
> document

However, that information only applies only to visibilitychange and not
to visibilityState — so it’s also not accurate to keep the
visibilityState support marked as partial.

Therefore, this change drops the partial_implementation flag, as well as
dropping the note entirely.