Update Firefox data for CSSKeyframesRule API (#6909)
This commit updates the Firefox data for the CSSKeyframesRule API
based upon manual testing.  The prefixed version was removed in the
same version it became unprefixed.  Additionally, this fixes the
capitalization in the prefix, and sets version numbers for all of the
subfeatures (also based upon manual testing, running
`[Moz]CSSKeyframesRule.prototype.[subfeature]` in console and checking
if it's defined).