Remove notes about document.all being readonly (#7012)
These notes were added in https://github.com/mdn/browser-compat-data/pull/3391,
after the change in https://chromium-review.googlesource.com/c/chromium/src/+/823513/.

Previously it was possible to assign document.all, and now it is not.
The uses of document.all that currently work also work with older
browsers, so there isn't anything that web developers need to be aware
of here. The only time when it might have mattered was at the time of
the change, if someone depending on being able to use document.all to
store their own values, which in itself is an unlikely scenario.