Drop null-value browsers from CSS gap properties
This change drops all the remaining null-value entries from the data in
the css tree — which were in the data for the gap, column-gap, and
row-gap properties. (Those remaining values were for the qq_android,
uc_android, and uc_chinese_android browsers — which we don’t record on
for most other CSS properties anyway.)

With this change, the css tree has no remaining entries with null values.