Update Chrome/Opera data for SVG element APIs using mdn-bcd-collector (#7197)
This commit sets results (from true/null to a real value) for Chrome
and Opera for various SVG element APIs based upon results from the
mdn-bcd-collector project, followed by mirroring to Chrome Android,
Opera Android, Samsung Internet, and WebView.  Data comes from the
results from Chrome 1-85, as well as Opera 12.14, 15, and 70, on both
macOS and Windows.

There are several results that change WebView from `false` to `1` (and
one entry from `3` to `1`).  This is because anything that is in
Chrome 1 is _guaranteed_ to be in WebView for Android 1 (as stated by
Joe Medley).