Correct Chromium data for HTML element APIs (#7179)
This commit corrects the Chrome data for the HTML element APIs based
upon results from the mdn-bcd-collector project (using results from
Chrome 1-86 and Opera 12.14, 15, and 70).

Helpful note: I noticed that a number of data was set to "Chrome 43";
however, from my testing, I've found that Chrome 43 was a magic number
of when all properties are initialized on the object prototype.  Thus,
it's best to trust results that indicate Chrome 43 less.