Correct Chrome data for HTMLIFrameElement API (#7243)
This commit corrects the Chrome data for the HTML `<iframe>` element
API based upon results from the mdn-bcd-collector project (using
results from Chrome 1-86 and Opera 12.14, 15, and 70).  Most of the
data was set to "Chrome 43"; however, from my testing, I've found that
Chrome 43 was a magic number of when all properties are initialized on
the object prototype. (Thus, it's best not to trust results that
indicate Chrome 43 as much.)