Set real values for HTML element APIs for IE/Edge (#7246)
This commit updates the data for the HTML element APIs for Internet
Explorer and Microsoft Edge using the mdn-bcd-collector project (using
results from IE 5.5 to 11, and Edge 13 to 85), aided with some manual
adjustments.  There are a few changes to existing data for consistency
reasons.

Because there is no real way to differentiate whether an element is
initialized using the HTMLElement API or its own API until IE 9, the
version numbers for the APIs themselves are guesstimated based upon
the values from the subfeatures (or when not possible, when the
elements themselves were supported).