Update Firefox for android video autoplay support
Adding note stating the current firefox for android state regarding video autoplay. Which most likely will make the feature not work at all, for most if not all users.

It is safe to say that a developper using "caniuse.com" expects to know if the feature will "most likely work for a percentage of users" and on which platform ( and thus, answer to the question "can I use it ?").
In the case of firefox for android the answer is "it will most likely not work" so adding this note is the bare minimum.
Could maybe change status to "partial support" to make it evident in the graphs and stats.

There is no patch-note, or any ressource on the internet or from mozilla stating the change. Which might signify that it is a oversight (bug ?) from the mozilla dev team. But it is the current state of the browser.