Revert "Remove never-supported WebXR APIs (#6943)" (#7412)
The original removals are a bit complicated. The APIs that use the
dictionaries are supported, so it's a bit messy to remove these without
breaking the MDN docs for those APIs. Plus they're still in the
specification, so I think it merits more investigation about how to
remove this data.

This reverts commit cb1d7a226d9d7a014cba8b00ac1ded111c206082.