Mark navigator.hardwareConcurrency Safari support (#7392)
This change marks navigator.hardwareConcurrency unsupported in Safari.

https://trac.webkit.org/changeset/169017/webkit added the support, so per
https://trac.webkit.org/browser/webkit/tags/Safari-601.1.56/Source/WebCore/page/Navigator.idl#L40
that puts in the WebKit version which shipped in Safari 9 — because per
https://trac.webkit.org/browser/webkit/tags/Safari-538.35/Source/WebCore/page/Navigator.idl#L38
it did not ship in Safari 8.

However, it shipped behind a build (compile) flag — and that build flag
was apparently never enabled by Apple for Safari releases.

https://trac.webkit.org/changeset/219372/webkit and
https://trac.webkit.org/changeset/219379/webkit eventually removed the
support from WebKit altogether. Per
https://trac.webkit.org/browser/webkit/releases/Apple/Safari%2011.0#WebCore/page
that means it was removed in Safari 11 — because per
https://trac.webkit.org/browser/webkit/releases/Apple/Safari%2010.0/WebCore/page
it was still present in the WebKit version that shipped with Safari 10.

But per https://github.com/mdn/browser-compat-data/pull/7393 and
https://github.com/mdn/browser-compat-data/pull/3752 we don’t mark
feature as supported if they require the use of a build/compile flag.
So this change instead simply marks it as version_added:false.