Update and correct Firefox data for Document API (#7494)
This commit updates the Firefox data for the Document API based upon
results from the mdn-bcd-collector project, along with a little bit of
mirroring to Firefox Android.  It appears that a number of features
were supported just a little bit earlier in Firefox than the data
proclaimed; this PR adds the correct data.