Correct Firefox data for HTMLContentElement API (#7501)
This commit corrects the Firefox data for the HTMLContentElement API
using results from the mdn-bcd-collector project. The collector
indicated that this API was supported by default without the need of
enabling the flag, but it was disabled by default in Firefox 52. While
the flag is present in Firefox 53, it was removed in Firefox 59, which
is more than two years ago, satisfying the flag removal requirement.