Correct Firefox data for HTMLShadowElement API (#7507)
This commit corrects the Firefox data for the HTMLShadowElement API
using results from the mdn-bcd-collector project. The collector
indicated that this API was supported by default without the need of
enabling the flag in Firefox 29 (not 28), but it was disabled by
default in Firefox 52. While the flag is present in Firefox 53, it was
removed in Firefox 58, which is more than two years ago, satisfying
the flag removal requirement.