Remove the api.Node.nodePrincipal entry
This is a Gecko "ChromeOnly" attribute not exposed to the web:
https://github.com/mozilla/gecko-dev/blob/8a4aa0c699d9ec281d1f576c9be1c6c1f289e4e7/dom/webidl/Node.webidl#L107

The string "nodePrincipal" does not appear in Chromium or WebKit source,
apart from one instane in a gecko_dom.js as part of a test suite.

The "API was moved" notes for Chromium are from
https://github.com/mdn/browser-compat-data/pull/1179 and must be a
mistake, as the string "nodePrincipal" has never appeared in a commit
message in the entire history of Chromium.

https://developer.mozilla.org/en-US/docs/Web/API/Node/nodePrincipal
should be redirected to Node.